Verify I2C is running at 400kHz with TWBR=12
The `i2c-SCL-TWBR_12.pdf` document verifies that the I2C is running at
400kHz when `TWBR=12`.  The `i2c-SCL-TWBR_default.pdf` verifies that the
I2C is running at 100kHz when `TWBR` is not set _(i.e., default)_.

__NB__ A 2.1kOhm pull-up resistor has been added to the `SCL` line to
decrease rise-time.  This helps to make the waveform closer to a
square-wave, as seen in the PDF diagram.

__NB__ The value of `TWBR` only needs to be set on the _master_, _i.e.,
the node that calls `Wire.beginTransmission()`/`Wire.endTransmission()`.